+++ src/constants/apiConstants.js
@@ -1,0 +1,6 @@
+const baseURIs = {
+    distance24: "https://www.distance24.org/",
+    apiStack: "http://api.ipstack.com/"
+};
+
+module.exports = { baseURIs };
